Adds clean_uploaded_pdfs.py and crontab for RobotReviewer web.
