+++ Dockerfile
@@ -24,7 +24,7 @@ RUN useradd --create-home --home /var/lib/deploy deploy
 # install apt-get requirements
 ADD apt-requirements.txt /tmp/apt-requirements.txt
 RUN apt-get -qq update -y
-RUN xargs -a /tmp/apt-requirements.txt apt-get install -y --no-install-recommends && apt-get clean
+RUN xargs -a /tmp/apt-requirements.txt apt-get install -y --no-install-recommends && apt-get clean && rm -rf /etc/cron.*/*
 
 # Certs
 RUN mkdir -p /etc/pki/tls/certs && \
@@ -58,6 +58,7 @@ RUN mkdir -p /var/lib/deploy/robotreviewer/data
 ADD server.py /var/lib/deploy/
 ADD server_api.py /var/lib/deploy/
 ADD entrypoint.sh /var/lib/deploy/
+ADD crontab /etc/cron.d/crontab
 ADD robotreviewer /var/lib/deploy/robotreviewer
 RUN chown -R deploy.deploy /var/lib/deploy/robotreviewer
