+++ entrypoint.sh
@@ -11,9 +11,14 @@ case "$OPTS" in
 
 celery)
     echo "[entrypoint.sh] Starting Celery"
-    cd /var/lib/deploy/ && celery -A robotreviewer.ml_worker worker --loglevel=info --concurrency=1 --pool=solo
+    cd /var/lib/deploy/ && celery -A robotreviewer.ml_worker worker --loglevel=info --concurrency=1 --pool=solo --uid=1000
     ;;
 web)
+    echo "[entrypoint.sh] Starting background cleanup task"
+    export > /var/lib/deploy/cron.env
+    chmod 0644 /etc/cron.d/crontab
+    crontab /etc/cron.d/crontab
+    cron
     echo "[entrypoint.sh] Starting RobotReviewer Web server"
     cd /var/lib/deploy/ && gunicorn --worker-class gevent --workers $GUNICORN_WORKERS --timeout $GUNICORN_WORKER_TIMEOUT -b 0.0.0.0:5000 server:app
     ;;
@@ -36,11 +41,6 @@ api-dev)
     export FLASK_APP=server_api:flask_app
     export FLASK_ENV=development
     cd /var/lib/deploy/ && flask run --host 0.0.0.0 --port 5001 --eager-loading --no-reload
-    ;;
-test)
-    echo "[entrypoint.sh] Running RobotReviewer unit tests"
-    cd /var/lib/deploy/ && python -m unittest &
-    sleep infinity
     ;;
 *)
     if [ ! -z "$(which $1)" ]
